1. separate CompetitionEditForm into dedicated file. 2. remove unused code.
