+++ src/LuckyDrawBot/Controllers/CompetitionEditForm.cs
@@ -1,0 +1,11 @@
+﻿namespace LuckyDrawBot.Controllers
+{
+    public class CompetitionEditForm
+    {
+        public string Gift { get; set; }
+        public string GiftImageUrl { get; set; }
+        public int WinnerCount { get; set; }
+        public string PlannedDrawTimeLocalDate { get; set; }
+        public string PlannedDrawTimeLocalTime { get; set; }
+    }
+}
