+++ src/LuckyDrawBot/Infrastructure/HealthCheck/DependencyServiceHealthCheckBuilderExtensions.cs
