+++ src/LuckyDrawBot/Infrastructure/HostingEnvironmentExtensions.cs
