+++ src/LuckyDrawBot/Infrastructure/ServiceProviderExtensions.cs
@@ -81,6 +81,5 @@ namespace System
         {
             InvokeMethod(serviceProvider, (object)instance, methodName);
         }
-
     }
 }
